fixa modularisering
